AI Update - Search
Updated AI state machine to now have basic searching:
- When chasing the Player, if the Player gets too far, the AI will change to the Search state.
- When in the Search state, the AI will pick a random number of points near where the Player was last seen to "search".
- For now the AI does not take into account Player hiding or tall grass (not in game yet), but the basic functionality is there so should be fast to integrate, fingers crossed.

Also, SIKE DID YOU THINK THE PREVIOUS COMMIT WAS THE LAST UPDATE TO ASTAR.C?
- Added function to assign the position of a valid floor tile of a map to a given pointer, given an epicenter, as well as min max radii.

+++ gam100-sample/source/FSM.h
@@ -42,7 +42,7 @@ struct FSM
 	// Functionality Variables:
 	float waitTime, elapsedTime;	// Allows a State to be "held" in wait for awhile, either as a countdown, or to do something for waitTime amount of seconds.
 	float tileSize;					// Size of a single map tile. Set when creating this FSM.
-	float searchCount;
+	int searchCount;
 };
 
 /// <summary>
